update: move def checks from init to config
