+++ src/commands/init.py
@@ -1,4 +1,3 @@
-import subprocess
 from datetime import datetime
 from colorama import Fore, Style, init
 
@@ -7,53 +6,12 @@ try:
 except ImportError:
     from .config import *
 
-def check_git_branch() -> str:
-    try:
-        branch_name = subprocess.check_output(['git', 'branch', '--show-current'],
-                                              stderr=subprocess.STDOUT).strip().decode('utf-8')
-        return branch_name
-    except subprocess.CalledProcessError as e:
-        print(Fore.RED + "Error: {e.output.decode('ufw-8')}")
-        return None
-        
-def check_git_origin() -> str:
-    try:
-        origin = subprocess.check_output(['git', 'remote', 'get-url', 'origin'],
-                                         stderr=subprocess.STDOUT).strip().decode('utf-8')
-        return origin
-    except subprocess.CalledProcessError as e:
-        print(Fore.RED + "Error: {e.output.deocde('utf-8')}")
-        return None
-
-#This is for the api check, if its GitHub oder GitLab
-def check_git_api_type():
-    #I should find a way to check if the ssh url(repo url) belongs to a GitHub Api or GitLab Api.
-    #Idk how to do it.
-    print("this api is gitlab? github? who know?")
-
-def check_git_username() -> str:
-    try:
-        origin = subprocess.check_output(['git', 'config', 'user.name'],
-                                         stderr=subprocess.STDOUT).strip().decode('utf-8')
-        return origin
-    except subprocess.CalledProcessError as e:
-        print(Fore.RED + "Error: {e.output.deocde('utf-8')}")
-        return None
-
-def check_git_email() -> str:
-    try:
-        origin = subprocess.check_output(['git', 'config', 'user.email'],
-                                         stderr=subprocess.STDOUT).strip().decode('utf-8')
-        return origin
-    except subprocess.CalledProcessError as e:
-        print(Fore.RED + "Error: {e.output.deocde('utf-8')}")
-        return None
 
 def init_git_config(args) -> dict:
 
     # Get Current date time
     currentDate = datetime.today().strftime('%Y-%m-%d %H:%M:%S')
-    
+
     branch = check_git_branch()
     origin = check_git_origin()
     username = check_git_username()
